pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.23.0",
    "gtts>=2.5.4",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pyaudio>=0.2.14",
    "sift-stack-py>=0.7.0",
    "speechrecognition>=3.14.3",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
]
